Start of Equanemone pendant board rev B!
Eff Eagle. I'm trying PCBmodE.

This repository should be placed in a boards/ directory which is next to
the pcbmode/ directory. From the level above that, run:

> python pcbmode/pcbmode.py -b equan-pendant -m
> python pcbmode/pcbmode.py -b equan-pendant --fab oshpark

Then you can compress everything created in
boards/equan-pendant/build/production and upload it to OSH Park!

This is just a basic starter board with the proper outline and not much
else. I brought the LED component over from another project just so this
project had some drills; OSH Park complains if there are no drills.